+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/RandomDotOrgAnonymousClientLiveTest.java
@@ -1,7 +1,16 @@
 package io.github.pr0methean.betterrandom.seed;
 
+import java.net.Proxy;
+import javax.annotation.Nullable;
+import javax.net.ssl.SSLSocketFactory;
+
 public class RandomDotOrgAnonymousClientLiveTest extends WebSeedClientLiveTest<RandomDotOrgAnonymousClient> {
   public RandomDotOrgAnonymousClientLiveTest() {
-    super(RandomDotOrgAnonymousClient.WITHOUT_DELAYED_RETRY);
+    super();
+  }
+
+  @Override protected RandomDotOrgAnonymousClient getSeedGenerator(@Nullable Proxy proxy,
+      @Nullable SSLSocketFactory socketFactory) {
+    return new RandomDotOrgAnonymousClient(proxy, socketFactory, false);
   }
 }